En ApiKeyView: Acceso a input. En apiKey.js: Pseudo-código setApiKey
