+++ test/apiKey.spec.js
@@ -1,0 +1,15 @@
+//import { getApiKey, setApiKey } from '../src/lib/apiKey.js';
+
+describe('getApiKey', () => {
+
+  it('debería devolver el valor de la API Key', () => {
+    // Desarrolla el test correspondiente aquí
+  });
+});
+
+describe('setApiKey', () => {
+
+  it('debería establecer correctamente la API Key', () => {
+    // Desarrolla el test correspondiente aquí
+  });
+});
